% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw-loess.R
\name{fw_loess}
\alias{fw_loess}
\title{Local regression with fixed width}
\usage{
fw_loess(
  data = NULL,
  formula,
  span,
  degree = 1,
  min_n = degree + 3,
  weights,
  weighting = TRUE
)
}
\arguments{
\item{data}{an optional data frame.}

\item{formula}{a formula specifying the numeric response and one numeric
predictor.}

\item{span}{width of the smoothing window.}

\item{degree}{degree of the polynomial.}

\item{min_n}{minimum number of points needed to perform local regression.}

\item{weights}{currently ignored.}

\item{weighting}{whether to use bicubic weighting or not.}
}
\description{
\link{lowess} it's awesome, but the span parameter is hard to interpret as the
actual points that are used depends on the size of the series. Furthermore,
the actual distance each point is to the centre of the regression is not
taken into account, which is a problem for series irregularly sampled.
}
